rename SearchCatalogQuery to BrowseCatalogQuery
