+++ Application/Catalog/ICatalogQueryRepository.cs
@@ -7,13 +7,13 @@
     using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
-    using SimpleMusicStore.Application.Catalog.Queries.Search;
+    using SimpleMusicStore.Application.Catalog.Queries.Browse;
 
     public interface ICatalogQueryRepository : IQueryRepository<MusicRecord>
     {
         Task<object> GetDetails(int id, CancellationToken cancellationToken = default);
         Task<object> GetBrowsingOptions(CancellationToken cancellationToken = default);
-        Task<IEnumerable<SearchCatalogMusicRecordOutputModel>> Search(
+        Task<IEnumerable<BrowseCatalogMusicRecordOutputModel>> Search(
                 Specification<MusicRecord> filterBy,
                 MusicRecordSorter sortBy,
                 int page,
